pyproject.toml
[project]
name = "claudesidian"
version = "0.1.0"
description = "An MCP server for integrating Obsidian with Claude"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "mcp>=1.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "httpx>=0.25.0",
    "questionary>=2.0.0",
    "uvicorn>=0.25.0",
    "watchdog>=3.0.0",  # For file system monitoring
    "networkx>=3.2.1",  # For graph operations
    "numpy>=1.26.0",    # For numerical operations
    "scikit-learn>=1.3.0",  # For text processing
    "html2text>=2020.1.16",  # For HTML to Markdown conversion
    "beautifulsoup4>=4.12.2",  # For HTML parsing
    "aiohttp>=3.9.1",  # For async HTTP requests
    "aiofiles>=23.2.1",  # For async file operations
    "readability-lxml>=0.8.1",  # For content extraction
    "pyppeteer>=1.0.2",  # For browser automation
    "fuzzywuzzy>=0.18.0",  # For fuzzy string matching
    "python-Levenshtein>=0.23.0",  # Optimizes fuzzywuzzy
]
requires-python = ">=3.9"

[project.scripts]
claudesidian = "claudesidian.__main__:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["claudesidian*"]
exclude = ["tests*"]

[project.urls]
Documentation = "https://github.com/yourusername/claudesidian/docs"
Source = "https://github.com/yourusername/claudesidian"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.6.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3